// Boardlight/src/api/generated/models/usersIndexUsersResponse.ts
/**
 * Generated by orval v6.17.0 🍺
 * Do not edit manually.
 * Blueboard
 * OpenAPI spec version: v4.0.0
 */
import type { UsersIndexUsersResponseUserGroup } from './usersIndexUsersResponseUserGroup';

export interface UsersIndexUsersResponse {
  id?: string;
  name?: string | null;
  email?: string | null;
  emailVerifiedAt?: string | null;
  realName?: string | null;
  class?: string | null;
  userGroups?: UsersIndexUsersResponseUserGroup[] | null;
  createdAt?: string;
  updatedAt?: string;
}